[frontend][api] Handle exception when changing column type
Fixes #2574

Altering column type fails when collation is not set. When database is
initialized for the first time, character set and collation optiosn are
set correctly. But in case a database has been upgraded since multiple
versions, then these options are sometimes not set.

This change will set character set and collation option for the column
if `change_column` raises a `StandardError` exception.