[frontend] Add validation for Issue.name
This patch is extension of PR#4128 in order to fix test failures.

Fixes #1215

[frontend] Validating issue.name against all regexes

Matching the name with combination of all regexes defined in each
IssueTracker and an additional regex to match numbers and hyphen.

rspec: Setting issue name to correct format. Replacing 'deleted' with
1234 as name of the deleted issue.

Also fixed `valid_name?` method to match the name with the regex defined
in the corresponding issue tracker.

Have added model tests (based on user_spec.rb) to validate simple
issues.